refactor modupdate package
